Created a popup folder for all the popup and moved the popup components into that folder
