--- frontend/src/app/components/events/participation/participation.component.html
@@ -1,1 +1,0 @@
-<p>participations works!</p>
