--- frontend/src/app/components/events/participation/participation.component.ts
@@ -1,15 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'app-participation',
-  templateUrl: './participation.component.html',
-  styleUrls: ['./participation.component.scss']
-})
-export class ParticipationComponent implements OnInit {
-
-  constructor() { }
-
-  ngOnInit(): void {
-  }
-
-}
