+++ frontend/src/app/popups/view-event/view-event.component.html
